+++ ignite/cmd/network_request_remove_validator.go
@@ -1,16 +1,11 @@
 package ignitecmd
 
 import (
-	"errors"
-	"fmt"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/spf13/cobra"
 
 	"github.com/ignite/cli/ignite/pkg/cliui"
 	"github.com/ignite/cli/ignite/pkg/cosmosutil"
 	"github.com/ignite/cli/ignite/services/network"
-	"github.com/ignite/cli/ignite/services/network/networkchain"
 	"github.com/ignite/cli/ignite/services/network/networktypes"
 )
 
@@ -57,40 +52,9 @@ func networkRequestRemoveValidatorHandler(cmd *cobra.Command, args []string) err
 		return err
 	}
 
-	chainLaunch, err := n.ChainLaunch(cmd.Context(), launchID)
-	if err != nil {
-		return err
-	}
-
-	c, err := nb.Chain(networkchain.SourceLaunch(chainLaunch))
-	if err != nil {
-		return err
-	}
-
-	var balance sdk.Coins
-	if c.IsAccountBalanceFixed() {
-		balance = c.AccountBalance()
-		if len(args) == 3 {
-			return fmt.Errorf(
-				"balance can't be provided, balance has been set by coordinator to %s",
-				balance.String(),
-			)
-		}
-	} else {
-		if len(args) < 3 {
-			return errors.New("account balance expected")
-		}
-		balanceStr := args[2]
-		balance, err = sdk.ParseCoinsNormalized(balanceStr)
-		if err != nil {
-			return err
-		}
-	}
-
-	return n.SendAccountRequest(
+	return n.SendValidatorRemoveRequest(
 		cmd.Context(),
 		launchID,
 		address,
-		balance,
 	)
 }
