+++ ignite/services/network/join.go
@@ -2,14 +2,10 @@ package network
 
 import (
 	"context"
-	"fmt"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	launchtypes "github.com/tendermint/spn/x/launch/types"
 
-	"github.com/ignite/cli/ignite/pkg/cliui/icons"
 	"github.com/ignite/cli/ignite/pkg/cosmosutil"
-	"github.com/ignite/cli/ignite/pkg/events"
 	"github.com/ignite/cli/ignite/pkg/xurl"
 	"github.com/ignite/cli/ignite/services/network/networkchain"
 	"github.com/ignite/cli/ignite/services/network/networktypes"
@@ -101,104 +97,4 @@ func (n Network) SendAccountRequestForCoordinator(ctx context.Context, launchID 
 	}
 
 	return n.SendAccountRequest(ctx, launchID, addr, amount)
-}
-
-// SendAccountRequest creates an add AddAccount request message.
-func (n Network) SendAccountRequest(
-	ctx context.Context,
-	launchID uint64,
-	address string,
-	amount sdk.Coins,
-) error {
-	addr, err := n.account.Address(networktypes.SPN)
-	if err != nil {
-		return err
-	}
-
-	msg := launchtypes.NewMsgSendRequest(
-		addr,
-		launchID,
-		launchtypes.NewGenesisAccount(
-			launchID,
-			address,
-			amount,
-		),
-	)
-
-	n.ev.Send("Broadcasting account transactions", events.ProgressStarted())
-
-	res, err := n.cosmos.BroadcastTx(ctx, n.account, msg)
-	if err != nil {
-		return err
-	}
-
-	var requestRes launchtypes.MsgSendRequestResponse
-	if err := res.Decode(&requestRes); err != nil {
-		return err
-	}
-
-	if requestRes.AutoApproved {
-		n.ev.Send(
-			"Account added to the network by the coordinator!",
-			events.Icon(icons.Bullet),
-			events.ProgressFinished(),
-		)
-	} else {
-		n.ev.Send(
-			fmt.Sprintf("Request %d to add account to the network has been submitted!", requestRes.RequestID),
-			events.Icon(icons.Bullet),
-			events.ProgressFinished(),
-		)
-	}
-	return nil
-}
-
-// SendValidatorRequest creates the RequestAddValidator message into the SPN
-func (n Network) SendValidatorRequest(
-	ctx context.Context,
-	launchID uint64,
-	peer launchtypes.Peer,
-	valAddress string,
-	gentx []byte,
-	gentxInfo cosmosutil.GentxInfo,
-) error {
-	addr, err := n.account.Address(networktypes.SPN)
-	if err != nil {
-		return err
-	}
-
-	msg := launchtypes.NewMsgSendRequest(
-		addr,
-		launchID,
-		launchtypes.NewGenesisValidator(
-			launchID,
-			valAddress,
-			gentx,
-			gentxInfo.PubKey,
-			gentxInfo.SelfDelegation,
-			peer,
-		),
-	)
-
-	n.ev.Send("Broadcasting validator transaction", events.ProgressStarted())
-
-	res, err := n.cosmos.BroadcastTx(ctx, n.account, msg)
-	if err != nil {
-		return err
-	}
-
-	var requestRes launchtypes.MsgSendRequestResponse
-	if err := res.Decode(&requestRes); err != nil {
-		return err
-	}
-
-	if requestRes.AutoApproved {
-		n.ev.Send("Validator added to the network by the coordinator!", events.ProgressFinished())
-	} else {
-		n.ev.Send(
-			fmt.Sprintf("Request %d to join the network as a validator has been submitted!", requestRes.RequestID),
-			events.ProgressFinished(),
-		)
-	}
-	return nil
 }
