+++ ignite/services/network/request.go
@@ -2,6 +2,9 @@ package network
 
 import (
 	"context"
+	"fmt"
+	"github.com/ignite/cli/ignite/pkg/cliui/icons"
+	"github.com/ignite/cli/ignite/pkg/cosmosutil"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	launchtypes "github.com/tendermint/spn/x/launch/types"
@@ -98,4 +101,148 @@ func (n Network) SubmitRequest(ctx context.Context, launchID uint64, reviewal ..
 
 	var requestRes launchtypes.MsgSettleRequestResponse
 	return res.Decode(&requestRes)
+}
+
+// SendAccountRequest creates an add AddAccount request message.
+func (n Network) SendAccountRequest(
+	ctx context.Context,
+	launchID uint64,
+	address string,
+	amount sdk.Coins,
+) error {
+	addr, err := n.account.Address(networktypes.SPN)
+	if err != nil {
+		return err
+	}
+
+	msg := launchtypes.NewMsgSendRequest(
+		addr,
+		launchID,
+		launchtypes.NewGenesisAccount(
+			launchID,
+			address,
+			amount,
+		),
+	)
+
+	n.ev.Send("Broadcasting account transactions", events.ProgressStarted())
+
+	res, err := n.cosmos.BroadcastTx(ctx, n.account, msg)
+	if err != nil {
+		return err
+	}
+
+	var requestRes launchtypes.MsgSendRequestResponse
+	if err := res.Decode(&requestRes); err != nil {
+		return err
+	}
+
+	if requestRes.AutoApproved {
+		n.ev.Send(
+			"Account added to the network by the coordinator!",
+			events.Icon(icons.Bullet),
+			events.ProgressFinished(),
+		)
+	} else {
+		n.ev.Send(
+			fmt.Sprintf("Request %d to add account to the network has been submitted!", requestRes.RequestID),
+			events.Icon(icons.Bullet),
+			events.ProgressFinished(),
+		)
+	}
+	return nil
+}
+
+// SendValidatorRequest creates the RequestAddValidator message into the SPN
+func (n Network) SendValidatorRequest(
+	ctx context.Context,
+	launchID uint64,
+	peer launchtypes.Peer,
+	valAddress string,
+	gentx []byte,
+	gentxInfo cosmosutil.GentxInfo,
+) error {
+	addr, err := n.account.Address(networktypes.SPN)
+	if err != nil {
+		return err
+	}
+
+	msg := launchtypes.NewMsgSendRequest(
+		addr,
+		launchID,
+		launchtypes.NewGenesisValidator(
+			launchID,
+			valAddress,
+			gentx,
+			gentxInfo.PubKey,
+			gentxInfo.SelfDelegation,
+			peer,
+		),
+	)
+
+	n.ev.Send("Broadcasting validator transaction", events.ProgressStarted())
+
+	res, err := n.cosmos.BroadcastTx(ctx, n.account, msg)
+	if err != nil {
+		return err
+	}
+
+	var requestRes launchtypes.MsgSendRequestResponse
+	if err := res.Decode(&requestRes); err != nil {
+		return err
+	}
+
+	if requestRes.AutoApproved {
+		n.ev.Send("Validator added to the network by the coordinator!", events.ProgressFinished())
+	} else {
+		n.ev.Send(
+			fmt.Sprintf("Request %d to join the network as a validator has been submitted!", requestRes.RequestID),
+			events.ProgressFinished(),
+		)
+	}
+	return nil
+}
+
+// SendValidatorRemoveRequest creates the RequestRemoveValidator message to SPN
+func (n Network) SendValidatorRemoveRequest(
+	ctx context.Context,
+	launchID uint64,
+	valAddress string,
+) error {
+	addr, err := n.account.Address(networktypes.SPN)
+	if err != nil {
+		return err
+	}
+
+	msg := launchtypes.NewMsgSendRequest(
+		addr,
+		launchID,
+		launchtypes.NewValidatorRemoval(
+			valAddress,
+		),
+	)
+
+	n.ev.Send("Broadcasting transaction", events.ProgressStarted())
+
+	res, err := n.cosmos.BroadcastTx(ctx, n.account, msg)
+	if err != nil {
+		return err
+	}
+
+	var requestRes launchtypes.MsgSendRequestResponse
+	if err := res.Decode(&requestRes); err != nil {
+		return err
+	}
+
+	if requestRes.AutoApproved {
+		n.ev.Send("Validator removed from network by the coordinator!", events.ProgressFinished())
+	} else {
+		n.ev.Send(
+			fmt.Sprintf(
+				"Request %d to remove validator from the network has been submitted!", requestRes.RequestID,
+			),
+			events.ProgressFinished(),
+		)
+	}
+	return nil
 }
